% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRBaseConvert.R
\name{getAllVersionInfo}
\alias{getAllVersionInfo}
\title{Get all miRBase version information}
\usage{
getAllVersionInfo()
}
\value{
The detailed version information is printed in the console for user reference.
}
\description{
This function return a reference for all miRBase versions' information including Version name,
Release date, miRNA number and Status.
}
\examples{
getAllVersionInfo()
}
\author{
Xu, Taosheng \email{taosheng.x@gmail.com}
}
